feat(wordlist): implement responsive word search pagination component that enhances the word list functionality with:
- Improved TypeScript type definitions and documentation
- Dynamic pagination based on viewport height
- Search filtering across multiple word fields
- Touch/swipe gesture support for navigation
- Automatic subscript numbering for duplicate words
