+++ src/Components/Organisms/Footer/ResourcesLink/ResourcesLink.d.tsx
@@ -1,3 +1,9 @@
+/**
+ * Defines the props for the ResourcesLink component.
+ * @param linkData - The data for the link.
+ * @param linkClasses - The CSS classes to apply to the link.
+ */
+
 export interface IResourceLinkProps {
   linkData: any,
   linkClasses: string | undefined
