+++ src/Components/Organisms/Wordlist/Wordlist.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { AppContentData } from '../../../Data/AppContent';
 import { BlockElement } from '../../UI/BlockLevel/BlockElement';
 import RenderElementHead from '../RenderElementHead/RenderElementHead';
-import WordlistObjects from './WordlistObjects/WordlistObjects';
+import WordlistObjects from './WordSearchPagination/WordSearchPagination';
 import { IWordlistProps } from './Wordlist.d';
 
 const {
